% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_adjustment_column.R
\name{add_adjustment_column}
\alias{add_adjustment_column}
\title{Add Adjustment Column}
\usage{
add_adjustment_column(cpi_table, zero_year, zero_month)
}
\arguments{
\item{cpi_table}{A table imported by \code{get_cpi()}.}

\item{zero_year}{The year to zero adjustment factors on.}

\item{zero_month}{The month to zero adjustment factors on.}
}
\value{
A CPI table with an additional column named pct_\emph{month}_\emph{year}.
}
\description{
Adds a column to a CPI table imported by \code{get_cpi()} containing adjustment
factors zeroed on a given month. The new column in named pct_\emph{year}_\emph{month}.
}
\seealso{
\code{\link[=get_cpi]{get_cpi()}}
}
