% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wage_column.R
\name{add_wage_column}
\alias{add_wage_column}
\title{Add Wage Column}
\usage{
add_wage_column(cpi_table, wage_table, zero_year, zero_month)
}
\arguments{
\item{cpi_table}{A table made by \code{add_adjustment_column()}.}

\item{wage_table}{A table with columns named \emph{year}, *\emph{month}, and \emph{salary}.}

\item{zero_year}{The year to zero wage on.}

\item{zero_month}{The month to zero wage on.}
}
\value{
A CPI table with an additional column named wage_\emph{year}_\emph{month}.
}
\description{
Adds a column to a CPI table that has had adjustment columns created by
\code{add_adjustment_column()} containing wages zeroed on a given month. The new
column in named wage_\emph{year}_\emph{month}.
}
\seealso{
\code{\link[=add_adjustment_column]{add_adjustment_column()}}
}
