% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_inflation.R
\name{adjust_for_inflation}
\alias{adjust_for_inflation}
\title{Adjust Wages for Inflation}
\usage{
adjust_for_inflation(wage_table, cpi_table)
}
\arguments{
\item{wage_table}{A table with \emph{year}, \emph{month}, and \emph{salary} columns.}

\item{cpi_table}{A table imported by \code{get_cpi()}.}
}
\value{
A tibble with pct_\emph{year}\emph{\emph{month} and wage}\emph{year}_\emph{month} added for
each month in \code{wage_table}.
}
\description{
Calculate percentual change and absolute change values for wages.
}
\details{
Given a tibble of wages, formatted with \emph{year}, \emph{month}, and \emph{salary}
columns, and a tibble containing CPI data imported by \code{get_cpi()}, the
function calculates relative and absolute inflation adjustments.

The function returns a tibble with columns named pct_\emph{year}\emph{\emph{month} and
wage}\emph{year}_\emph{month} for each month in the wage table.
}
\seealso{
\code{\link[=get_cpi]{get_cpi()}}
}
