% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_and_plot_adjusted_salary.R
\name{calculate_and_plot_adjusted_salary}
\alias{calculate_and_plot_adjusted_salary}
\title{Calculate and Plot Adjusted Salary}
\usage{
calculate_and_plot_adjusted_salary(
  wage_table,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{wage_table}{A tibble with *year*, *month*, and *salary* columns.}

\item{start_date}{Optional. A date in ISO format (yyyy-mm-dd).}

\item{end_date}{Optional. A date in ISO format (yyyy-mm-dd).}
}
\value{
A ggplot2 object.
}
\description{
Plots wages adjusted for inflation.
}
\details{
Takes a tibble of wages, formatted with columns *year*, *month*, and *salary*. Desired start and end dates for the plot can be specified in ISO format (yyyy-mm-dd). If no start and end dates are specified the whole period is plotted.

Returns a ggplot2 object.
}
