% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpi.R
\name{get_cpi}
\alias{get_cpi}
\title{Get Consumer Price Index}
\usage{
get_cpi()
}
\value{
A tibble with three columns; year, month, and value.
}
\description{
Downloads Consumer Price Index data from Statistics Sweden.
}
\details{
Downloads \href{http://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__PR__PR0101__PR0101A/KPIFastM2}{Swedish Consumer Price Index data from Statistics Sweden}.
The data set starts in January of 2006, with an index value of 100 occuring during 1980.

The function returns a tibble with three columns; year, month, and value. Months are numbered rather than named.
}
\examples{
cpi_data <- get_cpi()
}
