% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_value.R
\name{get_index_value}
\alias{get_index_value}
\title{Get Index Value}
\usage{
get_index_value(cpi_table, index_year, index_month)
}
\arguments{
\item{cpi_table}{A table imported by \code{get_cpi()}.}

\item{index_year}{The year to extract the index value for.}

\item{index_month}{The month to extract the index value for.}
}
\value{
A numeric index value.
}
\description{
Gets the index value from a table imported by \code{get_cpi()}.
}
\seealso{
\code{\link[=get_cpi]{get_cpi()}}
}
