% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wage.R
\name{get_wage}
\alias{get_wage}
\title{Get Wage}
\usage{
get_wage(wage_table, index_year, index_month)
}
\arguments{
\item{wage_table}{A table with columns named \emph{year}, *\emph{month}, and \emph{salary}.}

\item{index_year}{The year to extract the salary for.}

\item{index_month}{The month to extract the salary for.}
}
\value{
A salary.
}
\description{
Gets the salary from a table formatted with \emph{year}, \emph{month}, and \emph{salary} columns.
}
