% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inflation_adjusted_wage_table.R
\name{make_inflation_adjusted_wage_table}
\alias{make_inflation_adjusted_wage_table}
\title{Make Inflation Adjusted Wage Table}
\usage{
make_inflation_adjusted_wage_table(wage_table)
}
\arguments{
\item{wage_table}{A tibble of wages, formatted with \emph{date}, and \emph{salary}
columns. The day component of the dates is set to 1.}
}
\value{
A tibble with columns for \emph{date}, actual \emph{salary} and one inflation
adjuted column per past salary.
}
\description{
Given a tibble of wages, formatted with \emph{year}, \emph{month}, and \emph{salary}
columns, the function calculates the inflation adjusted wage for every past
salary up to the current day.
}
