% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentual_change.R
\name{percentual_change}
\alias{percentual_change}
\title{Calculate Percentual Change}
\usage{
percentual_change(start_value, end_value)
}
\arguments{
\item{start_value}{Start value.}

\item{end_value}{End value.}
}
\value{
Percentual change.
}
\description{
Calculates the percentual change from a start to an end value.
}
\examples{
percentual_change(140, 180)
}
