% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adjusted_wages.R
\name{plot_adjusted_wages}
\alias{plot_adjusted_wages}
\title{Plot Adjusted Wages}
\usage{
plot_adjusted_wages(adjusted_wage_table)
}
\arguments{
\item{adjusted_wage_table}{A tibble with wages adjusted for inflation.}
}
\value{
A ggplot2 plot.
}
\description{
Draw a plot from an adjusted wage table made by \code{make_inflation_adjusted_wage_table()}.
}
\seealso{
\code{\link[=make_inflation_adjusted_wage_table]{make_inflation_adjusted_wage_table()}}
}
